// api/src/types/apiContracts.ts
export interface WsRoleAlertPayload {
    guildId: string;
    guildName?: string;
    userId: string;
    userTag?: string;
    userAvatar?: string;
    isBot?: boolean;
    roleId: string;
    roleName?: string;
    action: "added" | "removed";
    isDangerous: boolean;
    permissions?: string[];
    executorInfo?: string;
    timestamp: number;
}